Introduce Plop for scaffolding
To ease the creation of new components, introduce the micro-generator
framework called [plop](https://www.npmjs.com/package/plop).

Introduce a new workspace in `./utils`. This is for common config and
tooling like babel, plop, etc. The first package is `@sens8/plop`
which contains the sens8 plopfile. The plopfile is where we set up the
generators for components, etc.